+++ pyppeteer/__init__.py
@@ -10,11 +10,12 @@ from appdirs import AppDirs
 
 __author__ = """Hiroyuki Takagi"""
 __email__ = 'miyako.dev@gmail.com'
-__version__ = '0.0.21'
-__chromimum_revision__ = '571375'
-__base_puppeteer_version__ = 'v1.2.0'
+__version__ = '0.0.22'
+__chromimum_revision__ = '575458'
+__base_puppeteer_version__ = 'v1.4.0'
 __pyppeteer_home__ = os.environ.get(
     'PYPPETEER_HOME', AppDirs('pyppeteer').user_data_dir)  # type: str
+DEBUG = False
 
 # Setup root logger
 _logger = logging.getLogger('pyppeteer')